requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.2
decorator==5.2.1
Django==4.2
environs==14.1.1
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
pycparser==2.22
python-dotenv==1.1.0
python-telegram-bot==13.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
